Critical fix for 64-bit architectures (problem reported by Thomas Leitner) : -    KIO_DATA << list.count(); +    KIO_DATA << (uint)list.count(); SlaveInterface uses a uint when decoding this.
svn path=/trunk/kdelibs/; revision=125646

+++ kio/slavebase.cpp
@@ -507,7 +507,7 @@ void SlaveBase::listEntry( const UDSEntry& entry, bool _ready )
 
 void SlaveBase::listEntries( const UDSEntryList& list )
 {
-    KIO_DATA << list.count();
+    KIO_DATA << (uint)list.count();
     UDSEntryListConstIterator it = list.begin();
     UDSEntryListConstIterator end = list.end();
     for (; it != end; ++it)
